requirements_v2.txt
# Backend Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
motor>=3.3.2
pymongo>=4.6.1
pandas>=2.2.2
numpy>=1.26.0
scikit-learn>=1.3.2
scipy>=1.11.4
python-multipart>=0.0.6
pydantic>=2.5.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
joblib>=1.3.2

# ML and Data Processing
statsmodels>=0.14.0
xgboost>=2.0.3
lightgbm>=4.1.0
optuna>=3.5.0
mlflow>=2.9.2

# WebSocket Support
websockets>=12.0

# File Processing
openpyxl>=3.1.2
xlrd>=2.0.1
python-calamine>=0.6.2
chardet>=5.2.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Security
cryptography>=41.0.8
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.2
black>=23.12.1
isort>=5.13.2
mypy>=1.8.0

# Frontend Dependencies (for reference)
# These should be installed via npm/yarn in the frontend directory
# next>=14.1.0
# react>=18.2.0
# @tanstack/react-query>=5.17.19
# framer-motion>=11.0.3
# three>=0.160.1
# @react-three/fiber>=8.15.16
# deck.gl>=9.0.16
# recharts>=2.10.4
# tailwindcss>=3.4.1